testing out ce loss with jitter

+++ pc_transforms.py
@@ -77,7 +77,7 @@ def jitter_point_cloud_torch(batch_data, sigma=0.01, clip=0.05):
     """
     T, N, C = batch_data.shape
     assert(clip > 0)
-    jittered_data = torch.clip(sigma * torch.randn(T, N, C, device=batch_data.device), -1*clip, clip)
+    jittered_data = torch.clip(sigma * torch.randn(T, N, C, device=batch_data.device, dtype=torch.float32), -1*clip, clip)
     jittered_data += batch_data
     return jittered_data
